Feature: Mean trip time joiner.
Changed the way in which queues start consuming (now its possible to consume from multiple queues in the same channel). Also added some waiting time to the client to avoid races in rabbit. The joiner is not done as we need to store the side table, send eof and send the matching trips to the aggregator