Feature: Complete Bully implementation with push healthchecks
Fixed bug in which the leader election algorithm would not start if a leader was already found. In case a new process joins with an id higher than the current leader, this process should take place and become the leader.

TODO: actually implement the "on_leader_callback" and "on_follower_callback"'s accordingly.

TODO 2: replace the push healthcheck model for a pull model