Feature: Watchdog Bully
The bully leader (main watcher) connects to every node that needs to be watched (including other watchers) and keeps track of the timestamps for each ping. If any node takes more than WATCHER_TIMEOUT_SECS, the leader considers that node fallen (no action is taken at the moment). The other watchers watch the leader and if the leader falls they start running another leader election. This implementation removes the burden of having to launch a thread for each host to be watched, but also comes with the price of static ip addressing (the watchers need to know the ip address of each of the hosts to be observed).